requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
plaid-python==9.1.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
python-multipart==0.0.6